# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/playlists": {
      "post": {
        "operationId": "PlaylistsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSmartPlaylistDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmartPlaylistDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new smart playlist",
        "tags": [
          "playlists"
        ]
      },
      "get": {
        "operationId": "PlaylistsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SmartPlaylistWithTracksDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all smart playlists",
        "tags": [
          "playlists"
        ]
      }
    },
    "/playlists/{id}": {
      "get": {
        "operationId": "PlaylistsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmartPlaylistWithTracksDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a smart playlist by ID",
        "tags": [
          "playlists"
        ]
      },
      "delete": {
        "operationId": "PlaylistsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a smart playlist",
        "tags": [
          "playlists"
        ]
      },
      "put": {
        "operationId": "PlaylistsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSmartPlaylistDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmartPlaylistDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a smart playlist",
        "tags": [
          "playlists"
        ]
      }
    },
    "/playlists/{id}/tracks": {
      "get": {
        "operationId": "PlaylistsController_getTracks",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination (minimum: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items per page (minimum: 1, maximum: 100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort tracks by. If not provided, uses the playlist default order.",
            "schema": {
              "enum": [
                "title",
                "artist",
                "album",
                "duration",
                "totalPlayCount",
                "lastPlayedAt",
                "rating",
                "addedAt"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort direction (ascending or descending)",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTracksDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get tracks for a smart playlist",
        "tags": [
          "playlists"
        ]
      }
    },
    "/playlists/{id}/tracks/play": {
      "get": {
        "operationId": "PlaylistsController_getTracksForPlay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shuffle",
            "required": false,
            "in": "query",
            "description": "Shuffle the tracks",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all track URIs for playing a playlist",
        "tags": [
          "playlists"
        ]
      }
    }
  },
  "info": {
    "title": "Libify API",
    "description": "Spotify library management API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "TrackSourceDto": {
        "type": "object",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "sourceId": {
            "type": "string",
            "description": "Spotify ID of the source"
          },
          "sourceName": {
            "type": "string",
            "description": "Name of the source (e.g., playlist name)"
          },
          "sourceType": {
            "type": "string",
            "description": "Type of source",
            "enum": [
              "LIKED_SONGS",
              "PLAYLIST",
              "ALBUM",
              "ARTIST_TOP_TRACKS"
            ]
          }
        },
        "required": [
          "createdAt",
          "id",
          "sourceType"
        ]
      },
      "TagDto": {
        "type": "object",
        "properties": {
          "color": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "TrackDto": {
        "type": "object",
        "properties": {
          "addedAt": {
            "format": "date-time",
            "type": "string"
          },
          "album": {
            "type": "string"
          },
          "albumArt": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "artistGenres": {
            "description": "Array of genre names from the track artist",
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "duration": {
            "type": "number",
            "description": "Duration in milliseconds"
          },
          "id": {
            "type": "string"
          },
          "lastPlayedAt": {
            "format": "date-time",
            "type": "string"
          },
          "ratedAt": {
            "format": "date-time",
            "type": "string"
          },
          "rating": {
            "type": "number",
            "maximum": 5,
            "minimum": 1
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrackSourceDto"
            }
          },
          "spotifyId": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagDto"
            }
          },
          "title": {
            "type": "string"
          },
          "totalPlayCount": {
            "type": "number"
          }
        },
        "required": [
          "addedAt",
          "artist",
          "artistGenres",
          "duration",
          "id",
          "sources",
          "spotifyId",
          "tags",
          "title",
          "totalPlayCount"
        ]
      },
...
      "PaginatedTracksDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "tracks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrackDto"
            }
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages",
          "tracks"
        ]
      },
...
      "PlaylistRuleDto": {
        "type": "object",
        "properties": {
          "daysValue": {
            "type": "number",
            "description": "For date rules - number of days"
          },
          "field": {
            "enum": [
              "album",
              "artist",
              "dateAdded",
              "duration",
              "lastPlayed",
              "playCount",
              "rating",
              "tag",
              "title"
            ],
            "type": "string"
          },
          "numberValue": {
            "type": "number"
          },
          "operator": {
            "enum": [
              "contains",
              "endsWith",
              "equals",
              "greaterThan",
              "hasAnyTag",
              "hasNoTags",
              "hasTag",
              "inLast",
              "isNotNull",
              "isNull",
              "lessThan",
              "notContains",
              "notEquals",
              "notHasTag",
              "notInLast",
              "startsWith"
            ],
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "operator"
        ]
      },
      "PlaylistCriteriaDto": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "number",
            "description": "Limit number of tracks"
          },
          "logic": {
            "enum": [
              "and",
              "or"
            ],
            "type": "string"
          },
          "orderBy": {
            "type": "string",
            "description": "Order by field"
          },
          "orderDirection": {
            "type": "object",
            "description": "Order direction"
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaylistRuleDto"
            }
          }
        },
        "required": [
          "logic",
          "rules"
        ]
      },
      "CreateSmartPlaylistDto": {
        "type": "object",
        "properties": {
          "criteria": {
            "$ref": "#/components/schemas/PlaylistCriteriaDto"
          },
          "description": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "criteria",
          "name"
        ]
      },
      "SmartPlaylistDto": {
        "type": "object",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "criteria": {
            "$ref": "#/components/schemas/PlaylistCriteriaDto"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "lastUpdated": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "createdAt",
          "criteria",
          "id",
          "isActive",
          "lastUpdated",
          "name"
        ]
      },
      "SmartPlaylistWithTracksDto": {
        "type": "object",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "criteria": {
            "$ref": "#/components/schemas/PlaylistCriteriaDto"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "lastUpdated": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "trackCount": {
            "type": "number"
          }
        },
        "required": [
          "createdAt",
          "criteria",
          "id",
          "isActive",
          "lastUpdated",
          "name",
          "trackCount"
        ]
      },
      "UpdateSmartPlaylistDto": {
        "type": "object",
        "properties": {
          "criteria": {
            "$ref": "#/components/schemas/PlaylistCriteriaDto"
          },
          "description": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        }
      }
    }
  }
}